The functionality of buying music resources has been implemented

+++ media-catalog-service/src/main/java/ru/mooncess/media_catalog_service/domain/MusicResourceInfo.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 import ru.mooncess.media_catalog_service.dto.AuthorInfo;
 
-import java.time.LocalDate;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -17,7 +17,7 @@ public class MusicResourceInfo {
     private String key;
     private int bpm;
     private String genre;
-    private double price;
+    private BigDecimal price;
     private String type;
     private List<AuthorInfo> authors;
 }
