+++ media-catalog-service/src/main/java/ru/mooncess/media_catalog_service/dto/AuthorInfo.java
@@ -3,10 +3,12 @@ package ru.mooncess.media_catalog_service.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class AuthorInfo {
     private String email;
     private String role;
-    private double percentageOfSale;
+    private BigDecimal percentageOfSale;
 }
